Fixed overlooked compile errors

+++ src/main/java/com/crowsofwar/avatar/client/gui/RadialMenu.java
@@ -116,7 +116,7 @@ public class RadialMenu extends Gui {
 		
 		AvatarPlayerData data = AvatarPlayerData.fetcher().fetchPerformance(mc.thePlayer);
 		if (data != null) {
-			int xp = data.getAbilityData(ability).getXp();
+			float xp = data.getAbilityData(ability).getXp();
 			String xpMessage = I18n.format("avatar.radial.xp", xp);
 			drawCenteredString(mc.fontRendererObj, xpMessage, x,
 					(int) (resolution.getScaledHeight() / 2 + mc.fontRendererObj.FONT_HEIGHT * 0.5),
